Support scalar pixel types.
Utilize NumericTraits::AssignToArray to copy a vector-like or scalar
pixel to an array. And use a "const reference" to the measurement
vector type so that either a copy will be created or it will just be a
reference to the value dependent of if they types match. That is for
scalar this type will be converted to a FixedArray<1>, which for
VariableLengthVectors this will just be a reference to the current
value.