Adding initial implementation of enforcing connectivity.
After the iterative clustering and labeling, each label is may not be
a single connected region. This optional step cleans up the results to
ensure that each label is a single connected region, and ensures each
region is of sufficient size.

This implementation is not ideal, and more efficient implementation
could be down with label maps if there was a efficient method to
convert scalar labels to label maps filter such that each isolated
label region gets a unique id. This implementation has two phases, the
first is multi-threaded, where for each cluster, the center pixel is
used to grow and make the region as labeled if it is of sufficient
size. It is anticipated that this is most of the pixels in the image
and relatively fast. The second step is single threaded, it scans the
whole image for unlabeled region, it labels them to a new Id if of
sufficient size, otherwise it labels it the previous valid label
encountered.